+++ editor/desktop/src/test/java/es/eucm/ead/editor/utils/GeoTester.java
@@ -49,6 +49,7 @@ import com.badlogic.gdx.graphics.glutils.ShapeRenderer;
 import com.badlogic.gdx.math.Polygon;
 import com.badlogic.gdx.math.Vector2;
 import com.badlogic.gdx.utils.Array;
+import es.eucm.ead.engine.utils.GeometryUtils;
 
 import java.util.Random;
