+++ editor/desktop/src/test/java/es/eucm/ead/editor/utils/ShapeEditor.java
@@ -51,6 +51,7 @@ import com.badlogic.gdx.utils.async.AsyncExecutor;
 import com.badlogic.gdx.utils.async.AsyncResult;
 import com.badlogic.gdx.utils.async.AsyncTask;
 import com.vividsolutions.jts.geom.Geometry;
+import es.eucm.ead.engine.utils.GeometryUtils;
 
 import java.util.concurrent.ConcurrentLinkedQueue;
